UefiPayloadPkg/PciHostBridgeLib: Return bootloader root bridge info when resource is not assigned
Today the PciHostBridgeLib returns NULL when resource is not assigned by bootloader.
It's to support certain platforms that the resource information cannot be determined
in bootloader phase when the resource is not assigned in bootloader.
Returning NULL allows the generic PciHostBridgeDxe driver in UPL exits and the platform
specific version in PPL starts.

The OVMF bootloader can determine the resource information.
This patch allows the generic PciHostBridgeDxe driver in UPL run and no need to
run the platform specific version in PPL.

Signed-off-by: Zhiguang Liu <zhiguang.liu@intel.com>